+++ rust/open-metadata/src/operations/mod.rs
@@ -1,8 +1,9 @@
-use crate::generated::Paging;
+use crate::generated::{CollectionDescriptor, Database, Paging, Table};
 pub use collections::*;
 pub use databases::*;
 pub use futures::StreamExt;
-use reqwest_pipeline::Continuable;
+use reqwest_pipeline::{collect_pinned_stream, Continuable, Response, Result as RPResult};
+use serde::Deserialize;
 pub use services::*;
 pub use tables::*;
 
@@ -33,5 +34,29 @@ impl<T> IntoIterator for PagedReturn<T> {
 
     fn into_iter(self) -> Self::IntoIter {
         self.data.into_iter()
+    }
+}
+
+impl PagedReturn<Database> {
+    pub(crate) async fn try_from(response: Response) -> RPResult<Self> {
+        let (_status_code, _headers, pinned_stream) = response.deconstruct();
+        let body = collect_pinned_stream(pinned_stream).await?;
+        Ok(serde_json::from_slice(&body).unwrap())
+    }
+}
+
+impl PagedReturn<Table> {
+    pub(crate) async fn try_from(response: Response) -> RPResult<Self> {
+        let (_status_code, _headers, pinned_stream) = response.deconstruct();
+        let body = collect_pinned_stream(pinned_stream).await?;
+        Ok(serde_json::from_slice(&body).unwrap())
+    }
+}
+
+impl PagedReturn<CollectionDescriptor> {
+    pub(crate) async fn try_from(response: Response) -> RPResult<Self> {
+        let (_status_code, _headers, pinned_stream) = response.deconstruct();
+        let body = collect_pinned_stream(pinned_stream).await?;
+        Ok(serde_json::from_slice(&body).unwrap())
     }
 }
